requirements.txt
flask
opencv-python-headless
numpy
Pillow
requests
werkzeug